Revert "sync main (#8)" (#10)
This reverts commit f0dd9a501ba5e4c1b14c6d4ebc68bd8f6b5c19ad.

+++ tests/pipelines/test_pipelines_image_to_text.py
@@ -51,7 +51,7 @@ class ImageToTextPipelineTests(unittest.TestCase):
         image_processor=None,
         feature_extractor=None,
         processor=None,
-        dtype="float32",
+        torch_dtype="float32",
     ):
         pipe = ImageToTextPipeline(
             model=model,
@@ -59,7 +59,7 @@ class ImageToTextPipelineTests(unittest.TestCase):
             feature_extractor=feature_extractor,
             image_processor=image_processor,
             processor=processor,
-            dtype=dtype,
+            torch_dtype=torch_dtype,
             max_new_tokens=20,
         )
         examples = [
